split service module
